move to neynar user folder
